RDM-8304: Refactor existing test file structure
    [RDM-8304](https://tools.hmcts.net/jira/browse/RDM-8304)

Move the following to the correct locations:
* white-black-list-roles-authorizer.spec.js
* delayed.js
