+++ test/spec/app/route/get-activities.spec.js
@@ -2,9 +2,7 @@ var config = require('config');
 var config = redis = ttlScoreGenerator = {};
 var activityService = require('../../../../app/service/activity-service')(config, redis, ttlScoreGenerator);
 var getActivitesRoute = require('../../../../app/routes/get-activities')(activityService, config);
-var addActivityRoute = require('../../../../app/routes/add-activity')(activityService, config);
 var httpMocks = require('node-mocks-http');
-var delayed = require('../../../../app/util/delayed');
 var chai = require("chai");
 var sinon = require("sinon");
 var sinonChai = require("sinon-chai");
